migrate: update testing configs and redistributes all `src/tests` files
- update tooling to point at new output paths
- move playwright and vscode tests to their respective (`packages/web/tests` and `apps/vscode-extension/tests/`) locations
