+++ apps/vscode-extension/tests/settingsFile.ts
