+++ apps/vscode-extension/tests/vscodeRegression.test.ts
