+++ package.json
@@ -140,10 +140,10 @@
     "pretest:e2e": "npm run compile && npm run compile-tests",
     "check-types": "tsc --noEmit",
     "lint": "eslint src",
-    "test": "node out/tests/runIntegrationTest.js",
+    "test": "node out/apps/vscode-extension/tests/runIntegrationTest.js",
     "test:all": "npm run test:pw && npm run test:vscode && npm run test:e2e",
-    "test:vscode": "node out/tests/runIntegrationTest.js --vscode",
-    "test:e2e": "node out/tests/runIntegrationTest.js --e2e",
+    "test:vscode": "node out/apps/vscode-extension/tests/runIntegrationTest.js --vscode",
+    "test:e2e": "node out/apps/vscode-extension/tests/runIntegrationTest.js --e2e",
     "test:pw": "npx playwright test",
     "test:pw:ui": "npm run compile && npm run compile-tests && npx playwright test --ui",
     "test:pw:headed": "npm run compile && npm run compile-tests && npx playwright test --headed",
