+++ packages/web/tests/globalTeardown.ts
@@ -5,8 +5,8 @@
  * This runs once after all test workers complete.
  */
 
-import { getWebServer } from '../../../packages/web/server/src/webServer';
-import * as logger from '../../../packages/core/src/logger';
+import { getWebServer } from '../server/src/webServer';
+import * as logger from '../../core/src/logger';
 
 async function globalTeardown(): Promise<void> {
     logger.info('[GlobalTeardown] Stopping MergeNB web server...');
