+++ packages/web/tests/syntaxHighlighting.spec.ts
@@ -5,7 +5,7 @@
  */
 
 import { test, expect } from './fixtures';
-import * as logger from '../../../packages/core/src/logger';
+import * as logger from '../../core/src/logger';
 import {
     readSettingsFileSnapshot,
     restoreSettingsFileSnapshot,
