--- src/tests/testHelpers.ts
@@ -1,214 +1,0 @@
-/**
- * @file testHelpers.ts
- * @description Shared utilities for MergeNB integration tests.
- * 
- * Contains common functions for:
- * - Server health checking
- * - Browser/page setup
- * - Cell source extraction
- * - Notebook structure validation
- */
-
-import * as http from 'http';
-import * as logger from '../../packages/core/src/logger';
-
-/** Shape returned by the /health endpoint */
-export interface HealthResponse {
-    status: string;
-    port: number;
-    activeSessions: number;
-    activeConnections: number;
-}
-
-/** A cell we expect to find on disk after resolution */
-export interface ExpectedCell {
-    rowIndex: number;
-    source: string;
-    cellType: string;
-    isConflict?: boolean;
-    isDeleted?: boolean;
-    metadata?: Record<string, unknown>;
-    hasOutputs?: boolean;
-    outputs?: Array<Record<string, unknown>>;
-    execution_count?: number | null;
-}
-
-/** Config written to disk by the runner, read by the test */
-export interface TestConfig {
-    workspacePath: string;
-    testName: string;
-    params?: any;
-}
-
-/** Check if the web server is up */
-export function checkHealth(port: number): Promise<boolean> {
-    return new Promise((resolve) => {
-        const url = `http://127.0.0.1:${port}/health`;
-        const req = http.get(url, { timeout: 500 }, (res) => {
-            res.resume(); // Drain body to free the socket
-            const isHealthy = res.statusCode === 200;
-            if (!isHealthy) {
-                logger.info(`[TestHelpers] Health check failed: got status ${res.statusCode} from ${url}`);
-            }
-            resolve(isHealthy);
-        });
-        req.on('error', (err) => {
-            logger.info(`[TestHelpers] Health check error on ${url}: ${err.message}`);
-            resolve(false);
-        });
-        req.on('timeout', () => {
-            logger.info(`[TestHelpers] Health check timeout on ${url}`);
-            req.destroy();
-            resolve(false);
-        });
-    });
-}
-
-/** Get detailed health info from the web server */
-export function getHealthInfo(port: number): Promise<HealthResponse | null> {
-    return new Promise((resolve) => {
-        const req = http.get(`http://127.0.0.1:${port}/health`, { timeout: 1000 }, (res) => {
-            if (res.statusCode !== 200) {
-                resolve(null);
-                return;
-            }
-            let data = '';
-            res.on('data', (chunk: string) => data += chunk);
-            res.on('end', () => {
-                try {
-                    resolve(JSON.parse(data));
-                } catch {
-                    resolve(null);
-                }
-            });
-        });
-        req.on('error', () => resolve(null));
-        req.on('timeout', () => { req.destroy(); resolve(null); });
-    });
-}
-
-/** Extract cell source as a string (handles both array and string formats) */
-export function getCellSource(cell: any): string {
-    if (!cell) return '';
-    return Array.isArray(cell.source) ? cell.source.join('') : (cell.source || '');
-}
-
-/** Parse a cell from data-cell attribute JSON */
-export function parseCellFromAttribute(cellJson: string | null, context: string): any {
-    if (!cellJson) {
-        logger.error(`Missing data-cell attribute for ${context}`);
-        throw new Error(`Missing data-cell attribute for ${context}`);
-    }
-    try {
-        return JSON.parse(decodeURIComponent(cellJson));
-    } catch (err) {
-        logger.error(`Failed to parse cell JSON for ${context}`, err);
-        throw new Error(`Failed to parse cell JSON for ${context}`);
-    }
-}
-
-/** Wait for the web server to start and return its port. Throws if not found within timeout. */
-export async function waitForServer(
-    getPort: () => Promise<number | undefined> | number | undefined,
-    timeoutMs = 30000
-): Promise<number> {
-    logger.info(`[TestHelpers] Waiting for server startup (timeout: ${timeoutMs}ms)`);
-    const maxAttempts = Math.ceil(timeoutMs / 500);
-    for (let i = 0; i < maxAttempts; i++) {
-        await new Promise(r => setTimeout(r, 500));
-        let serverPort = 0;
-        try {
-            const port = await Promise.resolve(getPort());
-            serverPort = typeof port === 'number' ? port : 0;
-        } catch (error) {
-            if (i % 10 === 0) {
-                logger.info(`[TestHelpers] Failed to query server port yet (attempt ${i + 1}/${maxAttempts}): ${error}`);
-            }
-            continue;
-        }
-        if (serverPort > 0) {
-            logger.info(`[TestHelpers] Found running server on port ${serverPort} (attempt ${i + 1}/${maxAttempts})`);
-            const isHealthy = await checkHealth(serverPort);
-            if (isHealthy) {
-                logger.info(`[TestHelpers] Server health check passed on port ${serverPort}`);
-                return serverPort;
-            }
-            if (i % 10 === 0) {
-                logger.info(`[TestHelpers] Server port present but health not ready yet: ${serverPort} (attempt ${i + 1}/${maxAttempts})`);
-            }
-        } else if (i % 10 === 0) {
-            logger.info(`[TestHelpers] Server not started yet (attempt ${i + 1}/${maxAttempts})`);
-        }
-    }
-    throw new Error('Web server did not start within timeout');
-}
-
-/** Wait for a session URL to be published by the extension test command. */
-export async function waitForSessionUrl(
-    getSessionUrl: () => Promise<string | undefined> | string | undefined,
-    timeoutMs = 15000
-): Promise<string> {
-    const maxAttempts = Math.ceil(timeoutMs / 500);
-    for (let i = 0; i < maxAttempts; i++) {
-        await new Promise(r => setTimeout(r, 500));
-        const raw = await Promise.resolve(getSessionUrl());
-        if (typeof raw !== 'string') continue;
-
-        let parsed: URL;
-        try {
-            parsed = new URL(raw);
-        } catch {
-            continue;
-        }
-
-        if (parsed.protocol !== 'http:' && parsed.protocol !== 'https:') continue;
-        if (!parsed.searchParams.has('session') || !parsed.searchParams.has('token')) continue;
-
-        return raw;
-    }
-    throw new Error('No valid session URL was created within timeout (requires http(s) with session & token params)');
-}
-
-/** Validate that a resolved notebook has valid .ipynb structure */
-export function validateNotebookStructure(notebook: any): void {
-    if (typeof notebook.nbformat !== 'number') {
-        throw new Error('Invalid notebook: missing nbformat');
-    }
-    if (typeof notebook.nbformat_minor !== 'number') {
-        throw new Error('Invalid notebook: missing nbformat_minor');
-    }
-    if (!notebook.metadata || typeof notebook.metadata !== 'object') {
-        throw new Error('Invalid notebook: missing metadata');
-    }
-    if (!Array.isArray(notebook.cells)) {
-        throw new Error('Invalid notebook: cells not an array');
-    }
-
-    for (let i = 0; i < notebook.cells.length; i++) {
-        const cell = notebook.cells[i];
-        if (!cell.cell_type) throw new Error(`Cell ${i}: missing cell_type`);
-        if (cell.source === undefined) throw new Error(`Cell ${i}: missing source`);
-        if (!cell.metadata) throw new Error(`Cell ${i}: missing metadata`);
-        if (cell.cell_type === 'code' && !Array.isArray(cell.outputs)) {
-            throw new Error(`Cell ${i}: code cell missing outputs`);
-        }
-    }
-}
-
-/**
- * Wait for the conflict file to be written (mtime within last 10 seconds).
- * Returns true if confirmed, false otherwise.
- */
-export async function waitForFileWrite(filePath: string, fs: typeof import('fs'), timeoutMs = 10000): Promise<boolean> {
-    const maxAttempts = Math.ceil(timeoutMs / 500);
-    for (let attempt = 0; attempt < maxAttempts; attempt++) {
-        await new Promise(r => setTimeout(r, 500));
-        try {
-            const stat = fs.statSync(filePath);
-            if (Date.now() - stat.mtimeMs < 10000) {
-                return true;
-            }
-        } catch { /* continue */ }
-    }
-    return false;
-}
